requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
streamlit==1.29.0

# Browser Automation
playwright==1.40.0
seleniumbase>=4.21.0  # Untuk bypass captcha/cloudflare dengan Playwright

# Image Handling
requests==2.31.0
Pillow==10.1.0

# Text Utilities
python-slugify==8.0.1

# Stability & Reliability
tenacity==8.2.3
fake-useragent==1.4.0

# Logging
loguru==0.7.2

# Event Loop Fix untuk Streamlit
nest-asyncio==1.6.0
